Initial values for log-likelihood optimization in rSPDE models with a latent stationary Gaussian Matern model

Auxiliar function to obtain domain-based initial values for log-likelihood optimization in rSPDE models with a latent stationary Gaussian Matern model

get.inital.values.rSPDE(
  mesh = NULL,
  mesh.range = NULL,
  include.nu = TRUE,
  log.scale = TRUE,
  nu_upper_bound = NULL
)

Arguments

mesh
An in INLA mesh

mesh.range
The range of the mesh.

include.nu
Should we also provide an initial guess for nu?

log.scale
Should the results be provided in log scale?

nu_upper_bound
Should an upper bound for nu be considered?

Value

A vector of the form (theta_1,theta_2,theta_3) or where theta_1 is the initial guess for tau, theta_2 is the initial guess for kappa and theta_3 is the initial guess for nu.