<<<<<<< HEAD

Data extraction from metric graphs for 'rSPDE' models

Extracts data from metric graphs to be used by 'INLA' and 'inlabru'.

Usage

graph_data_rspde(
  graph_rspde,
  name = "field",
  repl = NULL,
  repl_col = NULL,
  group = NULL,
  group_col = NULL,
  only_pred = FALSE,
  loc = NULL,
  bru = FALSE,
  tibble = FALSE,
  drop_na = FALSE,
  drop_all_na = TRUE
)

Arguments
<<<<<<< HEAD

graph_rspde
=======

graph_rspde
>>>>>>> a85650aaa7f5c4664408037a88cd875367a00eaa

An inla_metric_graph_spde object built with the rspde.metric_graph() function.

<<<<<<< HEAD
name

A character string with the base name of the effect.

repl
=======
name

A character string with the base name of the effect.

repl
>>>>>>> a85650aaa7f5c4664408037a88cd875367a00eaa

Which replicates? If there is no replicates, one can set repl to NULL. If one wants all replicates, then one sets to repl to .all.

<<<<<<< HEAD
repl_col

Which "column" of the data contains the replicate variable?

group
=======
repl_col

Which "column" of the data contains the replicate variable?

group
>>>>>>> a85650aaa7f5c4664408037a88cd875367a00eaa

Which groups? If there is no groups, one can set group to NULL. If one wants all groups, then one sets to group to .all.

<<<<<<< HEAD
group_col

Which "column" of the data contains the group variable?

only_pred

Should only return the data.frame to the prediction data?

loc

Locations. If not given, they will be chosen as the available locations on the metric graph internal dataset.

bru

Should the data be processed for inlabru?

tibble

Should the data be returned as a tidyr::tibble?

drop_na

Should the rows with at least one NA for one of the columns be removed? DEFAULT is FALSE. This option is turned to FALSE if only_pred is TRUE.

drop_all_na
=======
group_col

Which "column" of the data contains the group variable?

only_pred

Should only return the data.frame to the prediction data?

loc

Locations. If not given, they will be chosen as the available locations on the metric graph internal dataset.

bru

Should the data be processed for inlabru?

tibble

Should the data be returned as a tidyr::tibble?

drop_na

Should the rows with at least one NA for one of the columns be removed? DEFAULT is FALSE. This option is turned to FALSE if only_pred is TRUE.

drop_all_na
>>>>>>> a85650aaa7f5c4664408037a88cd875367a00eaa

Should the rows with all variables being NA be removed? DEFAULT is TRUE. This option is turned to FALSE if only_pred is TRUE.

Value
<<<<<<< HEAD
An 'INLA' and 'inlabru' friendly list with the data.

<<<<<<< HEAD ======= >>>>>>> a85650aaa7f5c4664408037a88cd875367a00eaa