Get the precision matrix of spacetimeobj objects

Function to get the precision matrix of a spacetimeobj object

Usage

# S3 method for class 'spacetimeobj'
precision(
  object,
  user_kappa = NULL,
  user_sigma = NULL,
  user_gamma = NULL,
  user_rho = NULL,
  ...
)

Arguments

object

The model object computed using spacetime.operators()

user_kappa

If non-null, update the range parameter of the covariance function.

user_sigma

If non-null, update the standard deviation of the covariance function.

user_gamma

If non-null, update the temporal range parameter of the covariance function.

user_rho

If non-null, update the drift parameter of the covariance function.

...

Currently not used.

Value

The precision matrix.

See also

Examples

s <- seq(from = 0, to = 20, length.out = 101)
t <- seq(from = 0, to = 20, length.out = 31)

op_cov <- spacetime.operators(space_loc = s, time_loc = t,
                             kappa = 5, sigma = 10, alpha = 1,
                             beta = 2, rho = 1, gamma = 0.05)
#> Error in spacetime.operators(space_loc = s, time_loc = t, kappa = 5, sigma = 10,     alpha = 1, beta = 2, rho = 1, gamma = 0.05): object 'mesh' not found
prec_matrix <- precision(op_cov)
#> Error: object 'op_cov' not found