Rational approximations of fractional intrinsic fields

rspde.intrinsic computes a Finite Element Method (FEM) approximation of a Gaussian random field defined as the solution to the stochastic partial differential equation (SPDE): $$(-\Delta)^{(\nu+d/2)/2}\tau u = W$$.

Usage

rspde.intrinsic(
  mesh,
  nu = NULL,
  nu.upper.bound = 2,
  mean.correction = FALSE,
  rspde.order = 1,
  prior.tau = NULL,
  prior.nu = NULL,
  prior.nu.dist = "lognormal",
  nu.prec.inc = 0.01,
  diagonal = 1e-05,
  type.rational.approx = "chebfun",
  shared_lib = "detect",
  debug = FALSE,
  ...
)

Arguments

mesh

Spatial mesh for the FEM approximation.

nu

If nu is set to a parameter, nu will be kept fixed and will not be estimated. If nu is NULL, it will be estimated.

nu.upper.bound

Upper bound for the smoothness parameter \(\nu\). If NULL, it will be set to 2.

mean.correction

Add mean correction for extreme value models?

rspde.order

The order of the covariance-based rational SPDE approach. The default order is 1.

prior.tau

A list specifying the prior for the variance parameter \(\tau\). This list may contain two elements: mean and/or precision, both of which must be numeric scalars.

prior.nu

a list containing the elements mean and prec for beta distribution, or loglocation and logscale for a truncated lognormal distribution. loglocation stands for the location parameter of the truncated lognormal distribution in the log scale. prec stands for the precision of a beta distribution. logscale stands for the scale of the truncated lognormal distribution on the log scale. Check details below.

prior.nu.dist

The distribution of the smoothness parameter. The current options are "beta" or "lognormal". The default is "lognormal".

nu.prec.inc

Amount to increase the precision in the beta prior distribution. Check details below.

diagonal

Number added to diagonal of Q for increased stability.

type.rational.approx

Which type of rational approximation should be used? The current types are "chebfun", "brasil" or "chebfunLB".

shared_lib

String specifying which shared library to use for the Cgeneric implementation. Options are "detect", "INLA", or "rSPDE". You may also specify the direct path to a .so (or .dll) file.

debug

Logical value indicating whether to enable INLA debug mode.

...

Additional arguments passed internally for configuration purposes.

Value

An object of class inla_rspde_intrinsic representing the FEM approximation of the intrinsic Gaussian random field.

Examples

library(fmesher)
n_loc <- 2000
loc_2d_mesh <- matrix(runif(n_loc * 2), n_loc, 2)
mesh_2d <- fm_mesh_2d(loc = loc_2d_mesh, cutoff = 0.03, max.edge = c(0.1, 0.5))
model <- rspde.intrinsic(mesh = mesh_2d)