Space-time random fields

spacetime.operators is used for computing a FEM approximation of a Gaussian random field defined as a solution to the SPDE $$d u + \gamma(\kappa^2 + \rho\cdot\Nabla - \Delta)^\alpha u = \sigma dW_C.$$ where C is a Whittle-Matern covariance operator with smoothness parameter \(\beta\) and range parameter \(\kappa\)

Usage

spacetime.operators(
  mesh_space = NULL,
  mesh_time = NULL,
  space_loc = NULL,
  time_loc = NULL,
  graph = NULL,
  kappa = NULL,
  sigma = NULL,
  gamma = NULL,
  rho = NULL,
  alpha = NULL,
  beta = NULL
)

Arguments

mesh_space

Spatial mesh for FEM approximation

mesh_time

Temporal mesh for FEM approximation

space_loc

Locations of mesh nodes for spatial mesh for 1d models.

time_loc

Locations of temporal mesh nodes.

graph

An optional metric_graph object. Replaces mesh for models on metric graphs.

kappa

Positive spatial range parameter

sigma

Positive variance parameter

gamma

Temporal range parameter.

rho

Drift parameter. Real number on metric graphs and one-dimensional spatial domains, a vector with two number on 2d domains.

alpha

Integer smoothness parameter alpha.

beta

Integer smoothness parameter beta.

Value

An object of type spacetimeobj.

Examples

s <- seq(from = 0, to = 20, length.out = 101)
t <- seq(from = 0, to = 20, length.out = 31)

op_cov <- spacetime.operators(space_loc = s, time_loc = t,
                             kappa = 5, sigma = 10, alpha = 1,
                             beta = 2, rho = 1, gamma = 0.05)
#> Error in spacetime.operators(space_loc = s, time_loc = t, kappa = 5, sigma = 10,     alpha = 1, beta = 2, rho = 1, gamma = 0.05): object 'mesh' not found
Q <- op_cov$Q
#> Error: object 'op_cov' not found
v <- rep(0,dim(Q)[1])
#> Error: object 'Q' not found
v[1565] <- 1
#> Error: object 'v' not found
Sigma <- solve(Q,v)
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'a' in selecting a method for function 'solve': object 'Q' not found

image(matrix(Sigma, nrow=length(s), ncol = length(t)))
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'x' in selecting a method for function 'image': object 'Sigma' not found