Update parameters of CBrSPDE objects

Function to change the parameters of a CBrSPDE object

# S3 method for CBrSPDEobj
update(
  object,
  user_nu = NULL,
  user_kappa = NULL,
  user_tau = NULL,
  user_m = NULL,
  ...
)

Arguments

object
The covariance-based rational SPDE approximation, computed using CBrSPDE.matern.operators

user_nu
If non-null, update the shape parameter of the covariance function.

user_kappa
If non-null, update the range parameter of the covariance function.

user_tau
If non-null, update the standard deviation of the covariance function.

user_m
If non-null, update the order of the rational approximation, which needs to be a positive integer.

...
Currently not used.

Value

It returns an object of class "CBrSPDEobj. This object contains the same quantities listed in the output of CBrSPDE.matern.operators.

See also

Examples

#Compute the covariance-based rational approximation of a 
#Gaussian process with a Matern covariance function on R
kappa <- 10
sigma <- 1
nu <- 0.8

#create mass and stiffness matrices for a FEM discretization
x <- seq(from = 0, to = 1, length.out = 101)
fem <- rSPDE.fem1d(x)

#compute rational approximation of covariance function at 0.5
tau <- sqrt(gamma(nu) / (sigma^2 * kappa^(2*nu) * (4*pi)^(1/2) * gamma(nu+1/2)))
op_cov <- CBrSPDE.matern.operators(C=fem$C, G=fem$G,nu=nu,
kappa=kappa,tau=tau,d=1,m=2)
op_cov
#> Type of approximation:  Covariance-Based Matern SPDE approximation 
#> Parameters of covariance function: kappa =  10 , tau =  0.0958753 , nu =  0.8 
#> Order or rational approximation:  2 
#> Size of discrete operators:  101  x  101 

#Update the range parameter of the model:
op_cov <- update(op_cov, user_kappa=20)
op_cov
#> Type of approximation:  Covariance-Based Matern SPDE approximation 
#> Parameters of covariance function: kappa =  20 , tau =  0.0958753 , nu =  0.8 
#> Order or rational approximation:  2 
#> Size of discrete operators:  101  x  101