rSPDE inlabru mapper

rSPDE inlabru mapper

Usage

# S3 method for inla_rspde
bru_get_mapper(model, ...)

# S3 method for bru_mapper_inla_rspde
ibm_n(mapper, ...)

# S3 method for bru_mapper_inla_rspde
ibm_values(mapper, ...)

# S3 method for bru_mapper_inla_rspde
ibm_jacobian(mapper, input, ...)

Arguments

model

An inla_rspde for which to construct or extract a mapper

...

Arguments passed on to other methods

mapper

A bru_mapper_inla_rspde object

input

The values for which to produce a mapping matrix

Examples

 #devel version
if (requireNamespace("INLA", quietly = TRUE) && 
     requireNamespace("inlabru", quietly = TRUE)){
library(INLA)
library(inlabru)

set.seed(123)
m <- 100
loc_2d_mesh <- matrix(runif(m * 2), m, 2)
mesh_2d <- inla.mesh.2d(
  loc = loc_2d_mesh,
  cutoff = 0.05,
  max.edge = c(0.1, 0.5)
)
sigma <- 1
range <- 0.2
nu <- 0.8
kappa <- sqrt(8 * nu) / range
op <- matern.operators(
  mesh = mesh_2d, nu = nu,
  range = range, sigma = sigma, m = 2,
  parameterization = "matern"
)
u <- simulate(op)
A <- inla.spde.make.A(
  mesh = mesh_2d,
  loc = loc_2d_mesh
)
sigma.e <- 0.1
y <- A %*% u + rnorm(m) * sigma.e
y <- as.vector(y)

data_df <- data.frame(y=y, x1 = loc_2d_mesh[,1],
                       x2 = loc_2d_mesh[,2])
coordinates(data_df) <- c("x1", "x2")
rspde_model <- rspde.matern(
  mesh = mesh_2d,
  nu_upper_bound = 2
)

cmp <- y ~ Intercept(1) + 
           field(coordinates, model = rspde_model)

rspde_fit <- bru(cmp, data = data_df)
summary(rspde_fit)
}
#> Loading required package: sp
#> This is INLA_24.03.29 built 2024-03-29 13:04:47 UTC.
#>  - See www.r-inla.org/contact-us for how to get help.
#>  - List available models/likelihoods/etc with inla.list.models()
#>  - Use inla.doc(<NAME>) to access documentation
#> Loading required package: fmesher
#> inlabru version: 2.10.1
#> INLA version: 24.03.29
#> Components:
#> Intercept: main = linear(1), group = exchangeable(1L), replicate = iid(1L)
#> field: main = cgeneric(coordinates), group = exchangeable(1L), replicate = iid(1L)
#> Likelihoods:
#>   Family: 'gaussian'
#>     Data class: 'SpatialPointsDataFrame'
#>     Predictor: y ~ .
#> Time used:
#>     Pre = 0.371, Running = 6.31, Post = 0.796, Total = 7.47 
#> Fixed effects:
#>            mean    sd 0.025quant 0.5quant 0.975quant  mode kld
#> Intercept 0.286 0.181     -0.073    0.286       0.65 0.286   0
#> 
#> Random effects:
#>   Name	  Model
#>     field CGeneric
#> 
#> Model hyperparameters:
#>                                           mean     sd 0.025quant 0.5quant
#> Precision for the Gaussian observations 168.37 87.296      49.97   151.18
#> Theta1 for field                         -3.31  0.465      -4.23    -3.31
#> Theta2 for field                          2.81  0.246       2.34     2.81
#> Theta3 for field                         -1.37  0.194      -1.75    -1.37
#>                                         0.975quant   mode
#> Precision for the Gaussian observations    383.805 118.08
#> Theta1 for field                            -2.404  -3.31
#> Theta2 for field                             3.314   2.79
#> Theta3 for field                            -0.988  -1.37
#> 
#> Deviance Information Criterion (DIC) ...............: -122.18
#> Deviance Information Criterion (DIC, saturated) ....: 199.41
#> Effective number of parameters .....................: 96.66
#> 
#> Watanabe-Akaike information criterion (WAIC) ...: -144.24
#> Effective number of parameters .................: 55.83
#> 
#> Marginal log-Likelihood:  -109.91 
#>  is computed 
#> Posterior summaries for the linear predictor and the fitted values are computed
#> (Posterior marginals needs also 'control.compute=list(return.marginals.predictor=TRUE)')
#> 
#devel.tag