Matern rSPDE model object for INLA

Creates an INLA object for a stationary Matern model with general smoothness parameter.

Usage

rspde.matern(
  mesh,
  nu_upper_bound = 4,
  rspde_order = 2,
  nu = NULL,
  sharp = TRUE,
  debug = FALSE,
  optimize = TRUE,
  prior.kappa = NULL,
  prior.nu = NULL,
  prior.tau = NULL,
  start.lkappa = NULL,
  start.nu = NULL,
  start.ltau = NULL,
  prior.nu.dist = c("beta", "lognormal"),
  nu.prec.inc = 1,
  type.rational.approx = c("chebfun", "brasil", "chebfunLB")
)

Arguments

mesh

The mesh to build the model. Should be an inla.mesh or an inla.mesh.1d object.

nu_upper_bound

Upper bound for the smoothness parameter.

rspde_order

The order of the covariance-based rational SPDE approach.

nu

If nu is set to a parameter, nu will be kept fixed and will not be estimated. If nu is NULL, it will be estimated.

sharp

The sparsity graph should have the correct sparsity (costs more to perform a sparsity analysis) or an upper bound for the sparsity? If TRUE, the graph will have the correct sparsity.

debug

INLA debug argument.

optimize

Should the model be optimized? In this case the sparsities of the matrices will be analyzed.

prior.kappa

a list containing the elements meanlog and sdlog, that is, the mean and standard deviation on the log scale.

prior.nu

a list containing the elements mean and prec for beta distribution, or loglocation and logscale for a truncated lognormal distribution. loglocation stands for the location parameter of the truncated lognormal distribution in the log scale. prec stands for the precision of a beta distribution. logscale stands for the scale of the truncated lognormal distribution on the log scale. Check details below.

prior.tau

a list containing the elements meanlog and sdlog, that is, the mean and standard deviation on the log scale.

start.lkappa

Starting value for log of kappa.

start.nu

Starting value for nu.

start.ltau

Starting value for log of tau.

prior.nu.dist

The distribution of the smoothness parameter. The current options are "beta" or "lognormal". The default is "beta".

nu.prec.inc

Amount to increase the precision in the beta prior distribution. Check details below.

type.rational.approx

Which type of rational approximation should be used? The current types are "chebfun", "brasil" or "chebfunLB".

Value

An INLA model.

Details

This function constructs a stationary Matern rSPDE model to be used with the INLA interface. The parameters are the range parameter \(\kappa\), the smoothness parameter \(\nu\) and the variance rescaling parameter \(\tau\).

For this model, an upper bound for the smoothness parameter \(\nu\) should be given. It is given by the nu_upper_bound argument.

It is very important to notice that the larger the value of nu_upper_bound the higher the computational cost to fit the model. So, it is generally best to initially fit a model with a small value of nu_upper_bound and increase it only if it is really needed (for instance, if the estimated smoothness parameter was very close to nu_upper_bound).

The following parameterization is used: $$\log(\tau) = \theta_1,$$ $$\log(\kappa) = \theta_2$$ and for \(\theta_3\) we can have a beta prior or a truncated lognormal prior distribution. In each case, the prior distribution has support on the interval \((0,\nu_{UB})\), where \(\nu_{UB}\) is nu_upper_bound. Then, the following parameterization is considered: $$\log\Big(\frac{\nu}{\nu_{UB}-\nu}\Big) = \theta_3.$$

By default, an optimized version of this model is considered. The optimized version is generally much faster for larger datasets, however it takes more time to build the model as the sparsity of the graph should be analyzed. However, for small datasets, it is possible that the time taken to analyze sparsity plus fitting the model is larger than the time taken to fit an unoptimized model. So, for a small dataset it might be convenient to set optimize=FALSE.

A way to use the optimized version but reduce the cost of sparsity analysis is to set sharp to FALSE. However, it should increase the cost of fitting the model. Therefore, one usually would not benefit from setting the sharp argument to FALSE when fitting the model to large datasets.

Finally, when considering a beta prior, the beta distribution will be parameterized in terms of its mean, say \(\mu\) and a precision parameter \(\phi\), which is such that the variance of the beta distribution is given by \(\mu(\nu_{UB}-\mu)/(1+\phi)\). The mean of the beta prior is determined by the prior.nu$mean, whereas the precision parameter is determined by the prior.nu$prec. If prior.nu$prec is NULL (which is the default case), the precision parameter is taken as $$\phi = \max\Big\{\frac{\nu_{UB}}{\mu}, \frac{\nu_{UB}}{\nu_{UB}-\mu}\Big\} + \textrm{nu.prec.inc},$$ where \(\mu\) is the prior mean of the smoothness parameter.

This choice of precision parameter is to ensure that the prior beta density has boundary values equal to zero (where the boundary values are defined either by continuity or by limits).

Hence, the higher the value of nu.prec.inc the more informative the prior is.

Examples

 #devel version
library(INLA)

#Organizing the data
data(PRprec)
data(PRborder)

Y <- rowMeans(PRprec[, 3 + 1:31])
ind <- !is.na(Y)
Y <- Y[ind]
coords <- as.matrix(PRprec[ind, 1:2])
alt <- PRprec$Altitude[ind]

seaDist <- apply(spDists(coords, PRborder[1034:1078, ], longlat = TRUE), 
1, min)

#Creating INLA mesh
prdomain <- inla.nonconvex.hull(coords, -0.03, -0.05, resolution = c(80, 80))
prmesh <- inla.mesh.2d(boundary = prdomain, max.edge = c(0.6, 1.2), cutoff = 0.3)

#Building the A matrix
Abar <- rspde.make.A(mesh = prmesh, loc = coords)

#Building the index
mesh.index <- rspde.make.index(name = "field", mesh = prmesh)

#Creating the model
rspde_model <- rspde.matern(mesh = prmesh)

#INLA stack
stk.dat <- inla.stack(data = list(y = Y), A = list(Abar, 1), tag = "est", 
  effects = list(c(mesh.index, 
                   list(Intercept = 1)), 
                   list(long = inla.group(coords[, 1]), 
                   lat = inla.group(coords[,2]),
                   seaDist = inla.group(seaDist))))

#INLA formula
f.s <- y ~ -1 + Intercept +  f(seaDist, model = "rw1") + 
f(field, model = rspde_model)  

#Fitting the model
rspde_fit <- inla(f.s, family = "Gamma", data = inla.stack.data(stk.dat), 
control.inla=list(int.strategy='eb'),
control.predictor = list(A = inla.stack.A(stk.dat)))

#The result
summary(rspde_fit)
#> 
#> Call:
#>    c("inla.core(formula = formula, family = family, contrasts = contrasts, 
#>    ", " data = data, quantiles = quantiles, E = E, offset = offset, ", " 
#>    scale = scale, weights = weights, Ntrials = Ntrials, strata = strata, 
#>    ", " lp.scale = lp.scale, link.covariates = link.covariates, verbose = 
#>    verbose, ", " lincomb = lincomb, selection = selection, control.compute 
#>    = control.compute, ", " control.predictor = control.predictor, 
#>    control.family = control.family, ", " control.inla = control.inla, 
#>    control.fixed = control.fixed, ", " control.mode = control.mode, 
#>    control.expert = control.expert, ", " control.hazard = control.hazard, 
#>    control.lincomb = control.lincomb, ", " control.update = 
#>    control.update, control.lp.scale = control.lp.scale, ", " 
#>    control.pardiso = control.pardiso, only.hyperparam = only.hyperparam, 
#>    ", " inla.call = inla.call, inla.arg = inla.arg, num.threads = 
#>    num.threads, ", " blas.num.threads = blas.num.threads, keep = keep, 
#>    working.directory = working.directory, ", " silent = silent, inla.mode 
#>    = inla.mode, safe = FALSE, debug = debug, ", " .parent.frame = 
#>    .parent.frame)") 
#> Time used:
#>     Pre = 4.93, Running = 24.1, Post = 0.0549, Total = 29.1 
#> Fixed effects:
#>           mean    sd 0.025quant 0.5quant 0.975quant mode   kld
#> Intercept 0.65 0.016      0.619     0.65       0.68   NA 0.008
#> 
#> Random effects:
#>   Name	  Model
#>     seaDist RW1 model
#>    field RGeneric2
#> 
#> Model hyperparameters:
#>                                                   mean       sd 0.025quant
#> Precision parameter for the Gamma observations   13.85    0.920     12.121
#> Precision for seaDist                          7710.45 6156.363   2056.109
#> Theta1 for field                                 -1.38    0.458     -2.312
#> Theta2 for field                                  1.40    0.263      0.877
#> Theta3 for field                                 -1.07    0.316     -1.674
#>                                                0.5quant 0.975quant    mode
#> Precision parameter for the Gamma observations    13.82     15.744   13.76
#> Precision for seaDist                           5957.79  24215.833 3891.59
#> Theta1 for field                                  -1.37     -0.499   -1.33
#> Theta2 for field                                   1.40      1.915    1.40
#> Theta3 for field                                  -1.07     -0.423   -1.10
#> 
#> Marginal log-Likelihood:  -1254.76 
#>  is computed 
#> Posterior summaries for the linear predictor and the fitted values are computed
#> (Posterior marginals needs also 'control.compute=list(return.marginals.predictor=TRUE)')
#> 
#devel.tag